Revert "LPD-65051 Move to internal.template.parser"
This reverts commit 977c18e73cc1541cac71f7f861d38409652aeeb3.
